Truely done

+++ src/Revenue-Expense/components/ListItem.js
@@ -3,8 +3,8 @@ import { useDispatch } from "react-redux";
 import { uiActions } from "../store/ui-slice";
 import { listActions } from "../store/list-slice";
 
-import editPencil from '../Image/edit_pencil.jpg';
-import deleteIcon from '../Image/deleteIcon.jpg';
+import editPencil from "../Image/edit_pencil.png";
+import deleteIcon from "../Image/deleteIcon.png";
 
 import classes from "./ListItem.module.css";
 
@@ -19,34 +19,39 @@ const ListItem = (props) => {
   };
 
   const deleteHandler = () => {
-    dispatch(listActions.removeList(order));
+    if (window.confirm("Do you really want to DELETE!!")) {
+      dispatch(listActions.removeList(order));
+    } else {
+      return;
+    }
   };
 
   return (
-    <tr className={classes.row}>
-      <td className={classes.category}>{category}</td>
-      <td>{date}</td>
-      <td>{order}</td>
-      <td>{person}</td>
-      <td>{detail}</td>
-      <td>{parseFloat(amount).toFixed(2)}</td>
-      <td>
+    <tr className="row">
+      <td className="col-2 justify-content-center">{category}</td>
+      <td className="col-2 justify-content-center">{date}</td>
+      <td className="col-1 justify-content-center">{order}</td>
+      <td className="col-1 justify-content-center">{person}</td>
+      <td className={`col-3 text-start ${classes.detail}`}>{detail}</td>
+      <td className="col-1 justify-content-center">
+        {parseFloat(amount).toFixed(2)}
+      </td>
+      <td className="col-1 p-2 justify-content-center">
         <input className={classes.trick} type="checkbox" />
       </td>
-      <td>
-        <img
-          className={classes.icon}
-          src={editPencil}
-          alt="Edit Pencil"
-          onClick={editHandler}
-        />
-
-        <img
-          className={classes.icon}
-          src={deleteIcon}
-          alt="Delete Icon"
-          onClick={deleteHandler}
-        />
+      <td className={`col-1 p-0 d-flex justify-content-around align-items-center ${classes.iconCover}`}>
+          <img
+            className={classes.icon}
+            src={editPencil}
+            alt="Edit Pencil"
+            onClick={editHandler}
+          />
+          <img
+            className={classes.icon}
+            src={deleteIcon}
+            alt="Delete Icon"
+            onClick={deleteHandler}
+          />
       </td>
     </tr>
   );
